✨ Add create-result-from-framework command and services
Introduce a new CreateResultFromFramework flow: command, handler and supporting services to create a result from a reporting framework. Adds services to create the result entity (including knowledge products), link the result to a ToC (and map levels), upsert ToC indicators and targets, and apply associations (contributors, projects, contributing centers). Adds small utils for standardized error throwing and safe hasOwn checks, wires new providers into the module and test setup, and refactors ResultsFrameworkReportingService to delegate creation to the new handler while also applying several minor internal cleanups (use of Set for IDs, safer status handling and type refinements).
