+++ onecgiar-pr-server/src/api/results-framework-reporting/commands/create-result-from-framework/create-result-from-framework.command.ts
@@ -1,0 +1,9 @@
+import { CreateResultsFrameworkResultDto } from '../../dto/create-results-framework.dto';
+import { TokenDto } from '../../../../shared/globalInterfaces/token.dto';
+
+export class CreateResultFromFrameworkCommand {
+  constructor(
+    public readonly payload: CreateResultsFrameworkResultDto,
+    public readonly user: TokenDto,
+  ) {}
+}
