+++ onecgiar-pr-server/src/api/results-framework-reporting/utils/reporting-framework-error.util.ts
@@ -1,0 +1,11 @@
+import { HttpStatus } from '@nestjs/common';
+
+export function throwReportingFrameworkError(
+  message: string,
+  status: HttpStatus = HttpStatus.BAD_REQUEST,
+): never {
+  const error = new Error(message);
+  (error as any).response = {};
+  (error as any).status = status;
+  throw error;
+}
