+++ onecgiar-pr-server/src/api/results-framework-reporting/utils/reporting-framework-object.util.ts
@@ -1,0 +1,7 @@
+type ObjectWithHasOwn = typeof Object & {
+  hasOwn(object: object, property: PropertyKey): boolean;
+};
+
+export function objectHasOwn(object: object, property: PropertyKey): boolean {
+  return (Object as ObjectWithHasOwn).hasOwn(object, property);
+}
